requirements.txt
cached-property==1.2.0
Jinja2==2.8
PyYAML==3.11
